Add backup functionality and prompt for whether to overwrite old remote repo contents

+++ repo_splitter/main.py
@@ -2,7 +2,7 @@ import shutil
 import os
 import tempfile
 import time
-from typing import Sequence
+from typing import Sequence, Optional
 
 import fire
 
@@ -18,7 +18,8 @@ from repo_splitter.github_tools.query import github_repo_from_clone_url
 
 def split_repo(repo_source: str, repo_dest: str, new_repo_name: str, keep_files: Sequence[str],
                github_token: str, all_branches: bool = False, include_tags: bool = False,
-               remove_files_from_old_repo: bool = True):
+               remove_files_from_old_repo: bool = True, keep_backup: bool = True,
+               auto_push_remove: bool = False, backup_dir: Optional[str] = None):
     """
     Splits an existing Git repository into two repositories by selecting which files should be
     split into a new one.
@@ -31,8 +32,18 @@ def split_repo(repo_source: str, repo_dest: str, new_repo_name: str, keep_files:
     :param all_branches: whether to include all branches in the new repo or only the primary (remote)/active (local) one
     :param include_tags: whether to keep tags from the old repo in the new one
     :param remove_files_from_old_repo: whether to remove the split files and history from the original repo
+    :param keep_backup: whether to retain a backup of the original repo in case something went wrong in removing history
+    :param auto_push_remove: pass True to avoid prompt for whether to push the original repo with history removed
+    :param backup_dir: pass file path to put backup of old repo there, otherwise uses repo_dest
     :return:
     """
+    if keep_backup:
+        if backup_dir is None:
+            backup_dir = os.path.join(repo_dest, 'backup')
+        if os.path.exists(backup_dir):
+            raise ValueError(f'backup folder {backup_dir} already exists. Remove it or pass keep_backup=False')
+
+
     # TODO: set up function to work with local only as well (github not required)
     with tempfile.TemporaryDirectory() as repo_temp_dest:
         print(f'Creating temporary repo from {repo_source}')
@@ -67,6 +78,9 @@ def split_repo(repo_source: str, repo_dest: str, new_repo_name: str, keep_files:
         print('Success')
         return
 
+    if keep_backup:
+        backup_repo = clone_repo(repo_source, backup_dir, all_branches=True)
+
     print(f'Cleaning up what was split off in the old repo')
     with tempfile.TemporaryDirectory() as repo_temp_dest:
         print(f'Cloning {repo_source} into temporary directory')
@@ -80,14 +94,26 @@ def split_repo(repo_source: str, repo_dest: str, new_repo_name: str, keep_files:
         print(f'Removing history in the original repo for files which were split off')
         remove_history_for_files_matching(repo, keep_files)
 
-        print('Pushing to remote for the original repo')
-        push_all_force(repo)
-
+        if not auto_push_remove:
+            print('Success. Please inspect the old repo to make sure nothing that was needed was removed.')
+            print('Once the history looks correct, enter Y to replace the remote repo contents')
+            print('If there is an issue with the history, enter N to exit')
+            push_repo_raw = input(f'Push modified history to {repo_source}? Y/N: ')
+            push_repo_str = push_repo_raw.strip().lower()[0]
+            push_repo = push_repo_str == 'y'
+        else:
+            print('auto_push_remove passed. Will automatically push to remote for original repo.')
+            push_repo = True
+        if push_repo:
+            print('Pushing to remote for the original repo')
+            push_all_force(repo)
+            print('If there is an issue, '
+                  'then you can go to your original local repo and run git push --all --force to reverse it')
+        else:
+            print('Not pushing modified history to original remote.')
         print('Removing temporary directory')
 
-    print('Success. Please inspect the old repo to make sure nothing that was needed was removed.')
-    print('If there is an issue, '
-          'then you can go to your original local repo and run git push --all --force to reverse it')
+
 
 
 def main():
